backend/requirements_doc_intel.txt
fastapi
uvicorn
pydantic
google-generativeai
google-cloud-texttospeech
python-dotenv
protobuf
sarvamai
